requirements.txt
Django==5.1.7
web3==7.10.0
eth-account==0.13.6
django-widget-tweaks==1.5.0
py-solc-x==2.0.3
eth-account==0.13.6
dotenv==0.9.9